Library reestructuration
