+++ src/dynamixel/controllers.hpp
@@ -1,0 +1,6 @@
+#ifndef DYNAMIXEL_CONTROLLERS_HPP_
+#define DYNAMIXEL_CONTROLLERS_HPP_
+
+#include "controllers/usb2dynamixel.hpp"
+
+#endif
