+++ src/dynamixel/dynamixel.hpp
@@ -1,11 +1,9 @@
 #ifndef DYNAMIXEL_HPP_
 #define DYNAMIXEL_HPP_
 
-#include "instruction_packet.hpp"
-#include "status_packet.hpp"
-#include "protocol1.hpp"
-#include "protocol2.hpp"
-#include "usb2dynamixel.hpp"
-#include "models.hpp"
+#include "controllers.hpp"
+#include "protocols.hpp"
+#include "servos.hpp"
+#include "errors.hpp"
 
 #endif
