+++ src/dynamixel/errors.hpp
@@ -1,0 +1,8 @@
+#ifndef DYNAMIXEL_ERRORS_HPP_
+#define DYNAMIXEL_ERRORS_HPP_
+
+#include "errors/error.hpp"
+#include "errors/crc_error.hpp"
+#include "errors/status_error.hpp"
+
+#endif
