--- src/dynamixel/models.hpp
@@ -1,19 +1,0 @@
-#ifndef DYNAMIXEL_MODELS_HPP_
-#define DYNAMIXEL_MODELS_HPP_
-
-#include "models/ax12.hpp"
-#include "models/ax12w.hpp"
-#include "models/ax18.hpp"
-#include "models/ex106.hpp"
-#include "models/mx12.hpp"
-#include "models/mx28.hpp"
-#include "models/mx64.hpp"
-#include "models/mx106.hpp"
-#include "models/xl320.hpp"
-#include "models/pro_l54_30_s400.hpp"
-#include "models/pro_l54_30_s500.hpp"
-#include "models/pro_l54_50_s290.hpp"
-#include "models/pro_l54_50_s500.hpp"
-#include "models/pro_l42_10_s300.hpp"
-
-#endif
