--- src/dynamixel/models/model_traits.hpp
@@ -1,12 +1,0 @@
-#ifndef DYNAMIXEL_MODELS_MODEL_TRAITS_HPP_
-#define DYNAMIXEL_MODELS_MODEL_TRAITS_HPP_
-
-namespace dynamixel {
-namespace models {
-    template<typename Model>
-    struct ModelTraits {        
-    };
-}
-}
-
-#endif
