+++ src/dynamixel/protocols.hpp
@@ -1,0 +1,7 @@
+#ifndef DYNAMIXEL_PROTOCOLS_HPP_
+#define DYNAMIXEL_PROTOCOLS_HPP_
+
+#include "protocols/protocol1.hpp"
+#include "protocols/protocol2.hpp"
+
+#endif
