+++ src/dynamixel/protocols/protocol2.hpp
@@ -1,15 +1,16 @@
-#ifndef DYNAMIXEL_PROTOCOL2_HPP_
-#define DYNAMIXEL_PROTOCOL2_HPP_
+#ifndef DYNAMIXEL_PROTOCOLS_PROTOCOL2_HPP_
+#define DYNAMIXEL_PROTOCOLS_PROTOCOL2_HPP_
 
 #include <stdint.h>
 #include <vector>
 #include <cassert>
 #include <boost/lexical_cast.hpp>
 
-#include "errors/crc_error.hpp"
-#include "errors/status_error.hpp"
+#include "../errors/crc_error.hpp"
+#include "../errors/status_error.hpp"
 
 namespace dynamixel {
+namespace protocols {
     class Protocol2 {
     public:
         typedef uint8_t id_t;
@@ -257,6 +258,7 @@ namespace dynamixel {
             return crc_accum;
         }
     };
+}
 }
 
 #endif
