+++ src/dynamixel/servos.hpp
@@ -1,0 +1,19 @@
+#ifndef DYNAMIXEL_SERVOS_HPP_
+#define DYNAMIXEL_SERVOS_HPP_
+
+#include "servos/ax12.hpp"
+#include "servos/ax12w.hpp"
+#include "servos/ax18.hpp"
+#include "servos/ex106.hpp"
+#include "servos/mx12.hpp"
+#include "servos/mx28.hpp"
+#include "servos/mx64.hpp"
+#include "servos/mx106.hpp"
+#include "servos/xl320.hpp"
+#include "servos/pro_l54_30_s400.hpp"
+#include "servos/pro_l54_30_s500.hpp"
+#include "servos/pro_l54_50_s290.hpp"
+#include "servos/pro_l54_50_s500.hpp"
+#include "servos/pro_l42_10_s300.hpp"
+
+#endif
