+++ src/dynamixel/servos/model_traits.hpp
@@ -1,0 +1,12 @@
+#ifndef DYNAMIXEL_SERVOS_MODEL_TRAITS_HPP_
+#define DYNAMIXEL_SERVOS_MODEL_TRAITS_HPP_
+
+namespace dynamixel {
+namespace servos {
+    template<typename Model>
+    struct ModelTraits {        
+    };
+}
+}
+
+#endif
