+++ src/dynamixel/servos/pro_l54_50_s290.hpp
@@ -1,23 +1,23 @@
-#ifndef DYNAMIXEL_PRO_L42_10_S300_HPP_
-#define DYNAMIXEL_PRO_L42_10_S300_HPP_
+#ifndef DYNAMIXEL_SERVOS_PRO_L54_50_S290_HPP_
+#define DYNAMIXEL_SERVOS_PRO_L54_50_S290_HPP_
 
 #include <stdint.h>
 
 #include "servo.hpp"
-#include "../protocol1.hpp"
+#include "../protocols/protocol1.hpp"
 
 namespace dynamixel {
-namespace models {
-    class ProL4210S300;
+namespace servos {
+    class ProL5450S290;
 
     template<>
-    struct ModelTraits<ProL4210S300> {
-        typedef Protocol2 protocol_t;
+    struct ModelTraits<ProL5450S290> {
+        typedef protocols::Protocol2 protocol_t;
 
         struct CT {
             static const protocol_t::address_t model_number = 0;
             typedef uint16_t model_number_t;
-            static const model_number_t model_number_value = 35072;
+            static const model_number_t model_number_value = 38176;
             static const protocol_t::address_t model_info = 2;
             typedef uint32_t model_info_t;
             static const protocol_t::address_t firmware_version = 6;
@@ -30,6 +30,8 @@ namespace models {
             typedef uint8_t return_delay_time_t;
             static const protocol_t::address_t operating_mode = 11;
             typedef uint8_t operating_mode_t;
+            static const protocol_t::address_t homing_offset = 13;
+            typedef uint32_t homing_offset_t;
             static const protocol_t::address_t moving_threshold = 17;
             typedef uint32_t moving_threshold_t;
             static const protocol_t::address_t highest_temperature_limit = 21;
@@ -70,8 +72,8 @@ namespace models {
             typedef uint16_t position_p_gain_t;
             static const protocol_t::address_t goal_position = 596;
             typedef int32_t goal_position_t;
-            static const goal_position_t min_goal_position = -2047;
-            static const goal_position_t max_goal_position = 2048;
+            static const goal_position_t min_goal_position = -180684;
+            static const goal_position_t max_goal_position = 180684;
             static const uint16_t min_goal_angle_deg = 0;
             static const uint16_t max_goal_angle_deg = 360;
             static const protocol_t::address_t moving_speed = 600;
@@ -101,15 +103,16 @@ namespace models {
         };
     };
 
-    class ProL4210S300 : public Servo<ProL4210S300> {
+    class ProL5450S290 : public Servo<ProL5450S290> {
     public:
-        typedef ProL4210S300 Model;
+        typedef ProL5450S290 Model;
 
-        ProL4210S300(typename protocol_t::id_t id) : Servo<ProL4210S300>(id) {};
+        ProL5450S290(typename protocol_t::id_t id) : Servo<ProL5450S290>(id) {};
 
         // Here we add the fields that are not common to all dynamixels
         READ_WRITE_FIELD(model_info);
         READ_WRITE_FIELD(operating_mode);
+        READ_WRITE_FIELD(homing_offset);
         READ_WRITE_FIELD(moving_threshold);
         READ_WRITE_FIELD(acceleration_limit);
         READ_WRITE_FIELD(torque_limit);
